move project into SpringWebApp
