+++ SpringWebApp/src/main/java/com/telusko/simpleWebApp/service/ProductLogService.java
